refactoring in a starbucks: - all buttons were moved to panelcontrols in their own function - the app modes were separated out of panel.js into their own respective components - that will be all.
