+++ src/components/EditMode.js
@@ -1,0 +1,33 @@
+const React = require('react');
+
+const PanelControls = require('./PanelControls');
+
+const EditMode = React.createClass({
+  propTypes: {
+    activeSnippet: React.PropTypes.object,
+    setMode:       React.PropTypes.func.isRequired
+  },
+
+  render() {
+    let activeSnippet = this.props.activeSnippet;
+    let text = null;
+
+    if ( activeSnippet ) {
+      text = activeSnippet.text
+    }
+
+    return(
+      <div className="panel-mode">
+        <div className="selected-mode">
+          <div className="edit-snippet-mode">
+            <textarea value={text}></textarea>
+          </div>
+       </div>
+
+       <PanelControls mode="edit" setMode={this.props.setMode} />
+      </div>
+    );
+  }
+});
+
+module.exports = EditMode;
